% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_equations_step_by_step.R
\name{step_by_step_eq}
\alias{step_by_step_eq}
\title{step_by_step_eq}
\usage{
step_by_step_eq(
  eqlist,
  before = "",
  after = "",
  title = " ",
  presentation_type = "quarto"
)
}
\arguments{
\item{eqlist}{The list of equations.  A list of equations in latex format.
The equations should be formatted as a character vector (see example below)
Backslashes need to be escaped and the results are otherwise printed asis.
Basically, this should just be a list of latex equations that are wrapped in
$$\begin{aligned}...\end{aligned}$$, so whatever is provided in this list is
just printed out.}

\item{before}{Optional text that goes before the equations.  This can be
formatted exactly the same way as they would be with stepping through the
equations line by line.}

\item{after}{Optional text that goes after the equations.}

\item{title}{Optional title for the slide.}

\item{presentation_type}{The type of presentation that the function is being
used for.  The default choice is `quarto`.  The other option is `xaringan`.}
}
\value{
NULL.  The function returns nothing but prints the output for
 generating slides with equation appearing one after another.
}
\description{
A function to allow display equations one line at a time in a revealjs
 presentation when the presentation is created using either quarto or
 xaringan.  The function ultimately creates individual slides for each
 equation and only increases the slide count for the first of these slides.
 The behavior mimics pausing in Beamer between lines of an equation.
}
\examples{
eqlist <- list("E[Y] &= E[Y|X=1]P(X=1) + E[Y|X=0]P(X=0) \\\\",
               "&= 1 \times 0.5 + 0 \times 0.5 \\\\",
               "&= 0.5")
before <- "Suppose that we know that $E[Y|X=1]=1$, $E[Y|X=0]=0$, $P(X=1)=0.5$,
 and $P(X=0) = 0.5$, then the law of iterated expecations implies that"
after <- "concluding text"
title <- "Law of Iterated Expectations Example"
step_by_step_eq(eqlist, before, after, title)

}
